Spread the love - the new allow/deny list base class is complete and implemented
